requirements.txt
pyspark==3.1.2
kafka-python
opencv-python
transformers
torch
streamlit
plotly
pandas
scipy
numpy
requests